管理后台商品控制器封装
